atualização nas classes DAO

+++ src/br/edu/ufersa/sei/model/VO/DiretorVO.java
@@ -3,7 +3,7 @@ package br.edu.ufersa.sei.model.VO;
 import br.edu.ufersa.sei.exception.InsertException;
 
 public class DiretorVO extends UsuarioVO{
-	private int idDir;
+	private long idDir;
 
 	//getters and setters
 	public long getIdDir() {
@@ -11,7 +11,7 @@ public class DiretorVO extends UsuarioVO{
 	}
 
 
-	public void setIdDir(int idDir) throws InsertException {
+	public void setIdDir(long idDir) throws InsertException {
 		if(idDir >= 0) {
 			this.idDir = idDir;
 		}else {
